refactor(userspace/falco): rename stats writer source files
