+++ userspace/falco/CMakeLists.txt
@@ -48,7 +48,7 @@ set(
   outputs_stdout.cpp
   outputs_syslog.cpp
   event_drops.cpp
-  statsfilewriter.cpp
+  stats_writer.cpp
   falco.cpp
 )
